requirements.txt
aiohttp==3.9.3
aiosignal==1.3.2
altgraph==0.17.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==25.3.0
bidict==0.23.1
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
multidict==6.4.3
packaging==24.2
propcache==0.3.1
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2025.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pyyaml==6.0.2
requests==2.31.0
simple-websocket==1.1.0
smmap==5.0.2
typing_extensions==4.13.2
websockets==12.0
wsproto==1.2.0
yarl==1.19.0
psutil==7.0.0
aiofiles==24.1.0
debugpy>=1.6
docker==7.1.0
distro==1.9.0
netifaces==0.11.0